Version 0.3.0
Instances are now created from a public base prototype (jQuery.plugin.base).
You can modify that prototype so any plugin you create can inherit
the same properties. This may be used to provide access to some of the
features in jQuery.ui.widget, in the case that you need them.

Upon initialization, the first argument will become the result of
merging the default plugin option with any the user may have passed in.
This differs from the old behavior which left the first argument intact
(so if the user passed in no options, it would be undefined).