Version 0.3.1
Fixes bug introduced in the last update. It was assumed that the first
argument passed in on initialization would always be an options object,
which would not be the case if a plugin method is called without
initializing first. The feature has been removed, options must be
referenced as this.options.

This also fixes a long-standing bug that merged non-objects into
options.